Use `QueryBuilder` for `Expression#to_sql`
I had hoped to actually break this into two commits, as the change to
`select_clause` is actually demonstrating a bug, but changing
`where_clause` to use the query builder, and changing `select_clause` to
use the expression both require each other.

I needed to have these functions take the query builder, as I could not
actually return a reference to the expression, as the type of the joined
query builders needs to be
`&(&Left::SelectClause, &Right::SelectClause)`, which cannot exist.

This also revealed to me that `Expressions::Bound` has the completely
incorrect trait bounds, as it should be `ToSql` not `ValuesToSql`. I
believe that I will be able to remove `ValuesToSql` entirely as this
refactoring continues.

The various `QuerySource` impls are getting more and more similar, which
means this is the right direction.