Move `select` to a generic context
This allows `SelectStatement` to implement `select`, without having to
duplicate implementations. It also allows us to remove
`SelectSqlQuerySource`, as `SelectStatement` now fills that role. I hope
to accomplish the same with the (currently commented out)
`FilteredQuerySource`.

The structure of this is a little bit funky, as it actually requires
higher kinded types, and we're faking it. What I'd much rather be able
to do is have a trait `SelectDsl` with no parameters, and have the type
`Output` be a type constructor taking a known number of aruguments.

The downside to having to implement it this way is that no helper
functions that depend on `select` can actually be part of `SelectDsl`.
`CountDsl` and `SelectSqlDsl` both have to be separate traits, even
though we know that anything which can implement any form of `SelectDsl`
can have the functions `count` and `select_sql`, as `CountStar` and
`SqlLiteral` impl `SelectableExpression` for all types.

Regardless, the language is what it is, so here we are. I have
absolutely no idea why this change caused two of my tests to start
spitting out the same error message 20 times. At the time of writing, I
am on a plane, but will be inquiring as soon as I can.

Additionally, the way that `SelectSqlDsl` is structured might seem odd.
It would make more sense for the `Self: SelectDsl<SqlLiteral<A>>` to be
a constraint on the trait itself, rather than every method inside of the
trait. However, type inference will *never* be able to infer the type
parameter for `select_sql`, so we do it this way, and impl the trait for
*every* type so that you're still able to do
`users.select_sql::<Int>("1 as one")` instead of
`SelectSqlDsl::<Int>::select_sql(users, "1 as one")`