Re-implement `filter` as a generic DSL
This will allow functions like `find` to take "a thing that can be
filtered with these types". Again, this would be much more
straightforward with HKT, where I can just take `FilterDsl`, and the
constraints for the predicate I'm attempting to pass (the actual
constraints for the predicate would be just as bad, but doing `T as
FilterDsl` would be less hard).

The type aliases are for my own sanity, as the expanded type signature
is really hairy here. I wonder if changing it from `find(table, pk)` to
`query_one(table.find(pk))` would clean things up at all...