Allow comparison against nullable columns
I think in the long run it will be more ergonomic to not have to use an
option in this case. If you want to do `WHERE column IS NULL`, you'll
need to write `filter(column.is_null())`. This means that if you
ultimately do want to take an optional value and filter on it, you'll
end up writing

```rust
match value {
    Some(expr) => column.eq(expr),
    None => column.is_null(),
}
```

In practice however, I've not seen this be a common enough case that I
think we need to support it. I could potentially be convinced to add a
separate `eq_option` function, though.

I need to see if it's possible to do this as a blanket impl without
specialization, or if there's some way to test that I've implemented
`AsExpression` for the nullable type sufficiently.

Fixes #4