Merge the implementations of `And` and `Eq` using a macro
After a little bit of thought, I think this will work for pretty much
any binary infix SQL operator. The only case I can come up with that
will need special handling will be `OR`, because I'll want to add
parenthesis, but I'll solve that in the `.or` helper rather than the
`Or` node.