Ensure simple queries are size 0 and don't have `WHERE 't' AND`
This gets us back to roughly the characteristics and semantics of before
I started this whole `SelectStatement` refactor. I couldn't decide
whether I wanted to test the `WHERE 't'` thing, since you would never
think to check for that unless you knew the specific implementation path
I took here. Since that glitch also meant that `users.as_query()` was of
size 1, which I do care about, I decided to test for that instead.

We're not using an enum or option for the where clause structs, because
we need a concrete output type for the filter DSL based on whether a
where clause was there before or not, which means they need to be
completely unrelated types.

It's worth noting that before this commit, the impl of `FilterDsl` for
`SelectStatement` never actually mentioned `SelectableExpression`, which
meant that the first implementation here passed the tests but allowed
incorrect queries. Luckily the compile fail tests that I added back in
the parent of this commit caught that error for me.