Implement missing comparison operators
This adds all simple comparison operators that were not yet implemented.
I've opted not to add bounds to the ordering operators, as nothing in
the PG docs implies that those operators are invalid for any types.

I'm not entirely sold on the method names for a lot of these, but I'm
currently using them as they're the method names of the equivalent
functions in the `cmp` module for Rust itself. However, I think I might
rename some of them to `not_eq`, `gt_eq`, and `lt_eq`.