Include OIDs when sending bind parameters
There are some cases where PG can't infer the type properly, such as
when an array is used inside of a bind parameter. This allows us to
disambiguate, and will allow writing a `ToSql<Array>` impl for various
types.

This is incomplete, and we should probably be including them in places
like insert, though I need to clean up that type heirarchy (the fact
that tuples needed to implement `oid` stands out as a smell to me).

I don't like that `NativeSqlType` is now coupled to PG, but I'm pretty
sure I can't decouple them without specialization, and it's not worth
trying until I have another target to find the right abstraction.

I do wonder if it'd make more sense to add `array_oid` to
`NativeSqlType`, instead of the giant if statement I have right now.
This would better match how PG represents it internally.