Minor public API tweaks
`yaqb::*` should not export `Result` or `Error` as this makes it a pain
to use.
`table::dsl::*` for two tables with no overlapping column names should
not overlap. We don't want `star` to be a top level namespace anyway.

I'll need to up the tuple length even further eventually, but compile
times grow dramatically as I increase this.