Add support for descending order
This type heirarchy is likely a bit loose. The fact that I've got an
expression of type `()`, and that I'm treating `Desc` as "Selectable"
are both smells to me that I'm missing an abstraction here. Need to
tighten this up, and test for things like not being able to pass
`name.desc()` to `select`