Add support for arbitrary user defined SQL functions
This macro will generate a type and constructor function for you. I've
had to require that the users name the arguments, as I need a name for
the type arguments. The generated type will follow the same rules as
everything else (is selectable if all its arguments are selectable, is
non-aggregate if its arguments are all non-aggregate, etc)

These compile-fail tests are proving really helpful. I almost skipped
one for this feature. I'm glad I didn't as it turned out I had written
constraints for `SelectableExpression` that were far too loose.