Add absolute minimal support for the `timestamp` type
It's not clear to me yet if there's any more useful type that we should
return here, but this covers the actual binary deserialization logic.
Once there is a concrete representation of a moment in time that exists
in the stdlib, we will certainly support that as well.

I chose to newtype this, as PG appears to represent time as number of
usec since 2000-01-01. That's pretty non-standard, so I'd prefer to be
explict on the meaning of the `i64`.