Add support for the `now` and `date` functions
I feel like the definition of `date` isn't going to be quite enough, as
I'm going to want to have a wider range of arguments be able to be
passed to it. That said, we don't have that need just yet, so we'll do
the simplest thing.

I didn't want to have to move these files around, but I can't seem to
use a macro from a sibling module, only a parent module (and only if the
`mod` declaration is after the macro)