rm `AsBindParam`
I'm attempting to clean up the `Insertable` family of traits as I
attempt to come up with an abstraction for `find_or_insert` and
`update_or_insert`. `AsBindParam` was a wart that has needed to go away
from some time now, so it's the easiest to graba first.

Ultimately `Expression` is absolutely not the right place for the
`to_insert_sql` function to live, but I do not think I can put it on
another trait cleanly until specialization lands.

As a side effect of this, pretty much everything is ending up behind a
reference in the insert expression, when it wasn't always before. I
think I missed some more cases, but this one can get a blanket impl once
specialization lands.