Generate more correct join SQL
The old generated code for `JoinTo` was a hold out from the very
earliest days of implementing, and is obviously incorrect. It's one of
the last hold-outs relying on methods like `qualified_name`, which are
also obviously incorrect (they cannot be correct without taking a
connection) which I would like to get rid of.

However, I can't actually create a node of `Eq<Serial, Integer>` right
now, as I'm requiring that they are the same type. I think eventually
I'd like to actually have a `ForeignKey` and `PrimaryKey` type to
represent these things, and make those comparable. It's not yet clear to
me if there are other cases where I need to allow equality of types
which are not the same, though.