Remove `Column#qualified_name`
This function is no longer used anywhere but tests. I also killed the
entire test that was using it, even though a few lines remained. This
test used to be more comprehensive, but at this point it's just testing
implementation details of trivially true statements where the entire
system would be broken if that test were failing.