Add a basic DSL for creating intervals for use as expressions
Note that in our tests, we are adding intervals to get intervals back,
not generating + in the sql query (though that will of course work for
expressions as well). This is still incomplete, as intervals can only be
added to other intervals (making them quite useless), and I also want to
support floats in this DSL as well.