Begin to rework the types to allow date + interval
I started to modify the impl of `std::ops::Add` to allow anything that
is addable on the Rhs, but I started to realize that I can't actually
write that, since the sql type would be an unused parameter of the impl.
This might be impossible to do "properly" without higher kinded types.

So I think that means we have to concede one of two things:

- `AsExpression` loses `SqlType` as a generic parameter, which means
  `String` can be usable as `VarChar` or `Text`, but not both.
- `AddableSqlType` loses `Rhs` as a generic parameter, and instead has
  it as an associated type. So types can only be added to exactly one
  other.

The second one might actualy be ok. Looking at the output of `SELECT *
FROM pg_operator WHERE oprname = '+'`, the only cases I see with the
same type on the left side more than once are for things like int2 +
int4, which I don't mind losing. I'm going to experiment in that
direction.