Add support for subselects
This is currently intended for use in `= ANY(subselect)`, but I'd
eventually like to add support to joining on subselects as well, as it's
often much more efficient, but that requires figuring out things like
aliasing, treating a subselect as a query source, etc etc.

I noticed that I was missing a constraint on `any`, as my tests passed
before I added the `SelectableExpression` impl for `SelectStatement`.
I've tightened that up, and added a test for it as well. I want to try
and figure out some way to automate testing that for new types.