Make `Expression` object safe
This is part of a refactoring to make boxed predicates possible to be
passed to `filter`. I've gone ahead and switched to dynamic dispatch for
all methods that take a query builder pre-emptively, as I'm going to
want to be able to box them as well.

Since this switches to dynamic dispatch, it will have a pretty
significant adverse impact of the performance of query construction,
even when unboxed. We can work around this by basically duplicating 100%
of our logic for all of the methods changed here, to have an object safe
and unsafe version, calling the unsafe version when possible. This will
have massive maintenance burdens however, and my code that actually
loads significant amounts of data is unaffected, as my select 0 rows
benchmarks are still measured in microseconds.

I can always add the static dispatch back at a later time.