Require the full type to be passed to `sql_function!`
I need to do this in most places that I'm automatically assuming
`yaqb::types::`. I've purposely designed the handling of various types
to be non-exhaustive and extensible, so other crates can add support for
additional types (for example, I'm working on a yaqb_full_text_search
crate).

As such, we can't assume that all types live in our module. I'm going to
hold off on making this change to macros which are intended to be called
from code gen based on a YAML file, as I have a lot of unanswered
questions there.