Add the ability to have columns which aren't selected by default
For example, crates.io has a column which is never selected, but is used
as an indexed ts_vector for queries updated via a trigger. This is also
commonly done via views. It makes sense, then, to not include this in
`table::all_columns`, etc, as it will never actually be in a select
statement, and we don't want it to be in the equivalent of
`table::star`.

That said, I'm not super happy about this API or the precident it sets,
but I want to put something in place here so I can move on.