Add support for joining through another table
I spent a bit of time thinking about how I might implement this as still
returning a predicate, but I couldn't quite come up with something that
I was happy with, so I went the easy route. The unfortunate side effect
here is that it's harder to compose these again, but we'll see if that
ends up being particularly painful.

As a side effect of this, you can't actually select anything from the
through column, but for the time being I'm fine with that. Once we
support multiple joins we'll need to figure out how to handle ambiguity
between multiple known ways to join, or maybe have this helper end up
just doing the same thing as manually joining through the through table.

It also occurs to me that something I haven't considered is making sure
that `users.left_outer_joins(posts).inner_joins(comments)` is either
returning `(User, Option<(Post, Comment)>)` or `(User, Option<Post>,
Option<Comment>)`.