Remove the `has_many` association method in favor of `belonging_to`
One of the most common problems you see in a lot of Rails apps which are
trying to avoid a god model is that no matter how hard you try, all of
your associations still end up having to go through the parent. While
the has_many will still be required if you want to try and select from a
join, if you're just loading the children of something, this will allow
the parent to have no idea.

The implementation is a bit of a mess at the moment, I'm planning on
refactoring out some of the shared code once I finish the rest of
`belongs_to`