Implement `Changeset` for `Vec<T>` and `Box<T>`
A valid use case has arisen in #26 that is impossible to work around at
the moment. We're going to eventually add a standard API to handle this,
so this implementation has no tests. It will get tested through that
API, once we've figured out what the API will look like. But I want to
get this through so it's possible to work around this by implementing
`AsChangeset` manually.