Get the README and root documentation a bit more ready for primetime
We're at the point where this can be judged for what it is. We don't
need the design explanation and "please help try it out" bit.

I also want to make sure that anyone who clicks documentation from
crates.io has an easy way to reach the getting started guide.