Change `insert` to mirror `update` and `delete`
This was a wart from the earliest days of implementation. It's
inconsistent, and feels weird in place with the rest of the commands.

I felt weird about the `Copy` bound on `InsertStatement`, but it ends up
being required and makes sense when you consider that we'll always be
passed references. However, it can potentially lead to some confusing
error messages.

The old methods on `Connection` are deprecated, and will be removed in
the next release.

Fixes #30.