Add support for the Postgres numeric type
This doesn't target any specific big decimal crate at the moment, it's
just minimal support for columns of that type. This was pretty
straightforward, the hardest part was really just getting an
implementation of `Arbitrary` which matches up with the behavior of PG's
reading.

The `max_digits = min(weight, scale)` is incorrect, but is a rough
approximation that will definitely get us below the length where it does
truncation. I don't want to duplicate that logic here, as it's complex.